requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
python-chess>=1.999
pydantic-settings>=2.4.0
pytest>=8.3.0
